backend/healix_app/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
boto3
opencv-python-headless
twilio
python-multipart
sqlalchemy
pydantic
pydantic-settings
apscheduler
pymysql
mysql-connector-python
cryptography
